Cache begin/end pass through event symbols when dlsyming them. Once we're in events disallowed/pass through, we won't be able to call our own dlsym to find them.
